Tools/chm_extractor/requirements.txt
# Python dependencies for CHM extraction
pychm>=0.8.6
